# src/za.co.grindrodbank.a3s/wwwroot/openapi-original.json
{
  "openapi" : "3.0.0",
  "info" : {
    "description" : "API Definition for the A3S. This service allows authentication, authorisation and accounting.",
    "title" : "A3S",
    "version" : "1.0.5"
  },
  "servers" : [ {
    "url" : "/"
  } ],
  "paths" : {
...
    "/functions" : {
      "get" : {
        "description" : "Search for functions.",
        "operationId" : "listFunctions",
        "parameters" : [ {
          "description" : "If this field is set, then the permission list is filled in\n",
          "explode" : true,
          "in" : "query",
          "name" : "permissions",
          "required" : false,
          "schema" : {
            "default" : false,
            "type" : "boolean"
          },
          "style" : "form"
        }, {
          "description" : "The page to view.",
          "explode" : true,
          "in" : "query",
          "name" : "page",
          "required" : false,
          "schema" : {
            "default" : 0,
            "format" : "int32",
            "minimum" : 0,
            "type" : "integer"
          },
          "style" : "form"
        }, {
          "description" : "The size of a page.",
          "explode" : true,
          "in" : "query",
          "name" : "size",
          "required" : false,
          "schema" : {
            "default" : 10,
            "format" : "int32",
            "maximum" : 20,
            "minimum" : 1,
            "type" : "integer"
          },
          "style" : "form"
        }, {
          "description" : "A search query filter on the description",
          "example" : "FilterByEntityType",
          "explode" : true,
          "in" : "query",
          "name" : "filterDescription",
          "required" : false,
          "schema" : {
            "maxLength" : 255,
            "minLength" : 0,
            "type" : "string"
          },
          "style" : "form"
        }, {
          "description" : "a comma separated list of fields in their sort order. Ascending order is assumed. Append desc after a field to indicate descending order.",
          "explode" : false,
          "in" : "query",
          "name" : "orderBy",
          "required" : false,
          "schema" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          },
          "style" : "form"
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/Function"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK",
            "headers" : {
              "X-Pagination" : {
                "description" : "Pagination header.",
                "explode" : false,
                "schema" : {
                  "$ref" : "#/components/schemas/PaginationHeaderResponse"
                },
                "style" : "simple"
              }
            }
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Bad Request."
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Not authenticated."
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Forbidden - You are not authorized to access the list of functions."
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Function list not found."
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "An unexpected error occurred."
          }
        },
        "summary" : "Search for functions.",
        "tags" : [ "Function" ]
      },
      "post" : {
        "description" : "Create a new function.",
        "operationId" : "createFunction",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/FunctionSubmit"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Function"
                }
              }
            },
            "description" : "OK."
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Bad Request."
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Not authenticated."
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Forbidden - You are not authorized to create functions."
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Function related entity (such as permissions) not found."
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "An unexpected error occurred."
          }
        },
        "summary" : "Create a new function.",
        "tags" : [ "Function" ]
      }
    },
    "/functions/{functionId}" : {
      "delete" : {
        "description" : "Deletes a function from A3S.",
        "operationId" : "deleteFunction",
        "parameters" : [ {
          "description" : "The UUID of the function to delete.",
          "explode" : false,
          "in" : "path",
          "name" : "functionId",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/FunctionId"
          },
          "style" : "simple"
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content."
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Invalid parameters."
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Not authenticated."
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Forbidden - Not authorized to delete functions."
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Function not found."
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "An unexpected error occurred."
          }
        },
        "summary" : "Delete a function.",
        "tags" : [ "Function" ]
      },
      "get" : {
        "description" : "Get a function by its UUID.",
        "operationId" : "getFunction",
        "parameters" : [ {
          "description" : "function",
          "explode" : false,
          "in" : "path",
          "name" : "functionId",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/FunctionId"
          },
          "style" : "simple"
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Function"
                }
              }
            },
            "description" : "OK."
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Bad Request."
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Not authenticated."
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Forbidden - You are not authorized to access the function."
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Function not found."
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "An unexpected error occurred."
          }
        },
        "summary" : "Get a function.",
        "tags" : [ "Function" ]
      },
      "put" : {
        "description" : "Update a function by its UUID.",
        "operationId" : "updateFunction",
        "parameters" : [ {
          "description" : "The UUID of the function.",
          "explode" : false,
          "in" : "path",
          "name" : "functionId",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/FunctionId"
          },
          "style" : "simple"
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/FunctionSubmit"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Function"
                }
              }
            },
            "description" : "OK"
          },
          "204" : {
            "description" : "No Content."
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Bad Request."
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Not authenticated."
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Forbidden - You are not authorized to update the function."
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "Functions not found."
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description" : "An unexpected error occurred."
          }
        },
        "summary" : "Update a function.",
        "tags" : [ "Function" ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "Permission" : {
        "description" : "Permission for an application\n",
        "example" : {
          "name" : "name",
          "description" : "description",
          "uuid" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91"
        },
        "properties" : {
          "uuid" : {
            "format" : "uuid",
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "FunctionId" : {
        "format" : "uuid",
        "type" : "string"
      },
      "Function" : {
        "description" : "A Function - A collection of permissions to denote a piece of functionality in an application. Must be assigned to an application. May optionally be associated with a sub-realm by specifying a 'subRealmId'.\n",
        "example" : {
          "subRealmId" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91",
          "permissions" : [ {
            "name" : "name",
            "description" : "description",
            "uuid" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91"
          }, {
            "name" : "name",
            "description" : "description",
            "uuid" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91"
          } ],
          "name" : "name",
          "description" : "description",
          "applicationId" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91",
          "uuid" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91"
        },
        "properties" : {
          "uuid" : {
            "format" : "uuid",
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "applicationId" : {
            "description" : "The UUID of an application.",
            "format" : "uuid",
            "type" : "string"
          },
          "subRealmId" : {
            "description" : "The UUID identifier for a sub-realm.",
            "format" : "uuid",
            "type" : "string"
          },
          "permissions" : {
            "items" : {
              "$ref" : "#/components/schemas/Permission"
            },
            "type" : "array"
          }
        },
        "required" : [ "applicationId", "description", "name", "permissions", "uuid" ],
        "type" : "object"
      },
      "FunctionSubmit" : {
        "description" : "Model used to create an new function or update an existing Function. May optionally be associated with a sub-realm by specifying a 'subRealmId'.\n",
        "example" : {
          "subRealmId" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91",
          "permissions" : [ "046b6c7f-0b8a-43b9-b35d-6489e6daee91", "046b6c7f-0b8a-43b9-b35d-6489e6daee91" ],
          "name" : "name",
          "description" : "description",
          "applicationId" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91",
          "uuid" : "046b6c7f-0b8a-43b9-b35d-6489e6daee91"
        },
        "properties" : {
          "uuid" : {
            "format" : "uuid",
            "type" : "string"
          },
          "name" : {
            "description" : "The name of the function.",
            "type" : "string"
          },
          "description" : {
            "description" : "A brief description of the function.",
            "type" : "string"
          },
          "applicationId" : {
            "description" : "The UUID of an application.",
            "format" : "uuid",
            "type" : "string"
          },
          "subRealmId" : {
            "description" : "The UUID identifier for a sub-realm.",
            "format" : "uuid",
            "type" : "string"
          },
          "permissions" : {
            "description" : "A list of permission UUIDs that are to be added to the function.",
            "items" : {
              "format" : "uuid",
              "type" : "string"
            },
            "type" : "array"
          }
        },
        "required" : [ "applicationId", "description", "name", "permissions", "uuid" ],
        "type" : "object"
      },
...
      "ErrorResponse" : {
        "description" : "Error response object to give more info with HTTP status code",
        "properties" : {
          "message" : {
            "type" : "string"
          }
        },
        "required" : [ "message" ],
        "type" : "object"
      },
      "PaginationHeaderResponse" : {
        "description" : "This is the standard response for requests that have paginated collections",
        "properties" : {
          "total" : {
            "description" : "The total number of pages in the result set",
            "format" : "int32",
            "minimum" : 1,
            "type" : "integer"
          },
          "count" : {
            "description" : "The total number of results in the result set",
            "format" : "int32",
            "minimum" : 1,
            "type" : "integer"
          },
          "size" : {
            "default" : 10,
            "description" : "The number items to include in a page of results. The page size for the result set",
            "format" : "int32",
            "maximum" : 20,
            "minimum" : 1,
            "type" : "integer"
          },
          "current" : {
            "description" : "The position of the page in the paged result set that is being returned",
            "format" : "int32",
            "minimum" : 1,
            "type" : "integer"
          },
          "first" : {
            "description" : "The link to the first page of results containing [size] results. This link includes the page, size, filter, orderBy and fields query parameters.",
            "format" : "url",
            "type" : "string"
          },
          "last" : {
            "description" : "The link to the last page of results containing [size] results. This link includes the page, size, filter, orderBy and fields query parameters. If this is the first page in the result set then this will be [null].",
            "format" : "url",
            "type" : "string"
          },
          "prev" : {
            "description" : "The link to the previous page of results containing [size] results. This link includes the page, size, filter, orderBy and fields query parameters. If this is the first page in the result set then this will be [null].",
            "format" : "url",
            "type" : "string"
          },
          "next" : {
            "description" : "The link to the next page of results containing [size] results. This link includes the page, size, filter, orderBy and fields query parameters. If this is the last page in the result set then this will be [null].",
            "format" : "url",
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
    }
  }
}